+++ pkg/util/metrics/BUILD
@@ -8,43 +8,20 @@ load(
 
 go_library(
     name = "go_default_library",
-    srcs = [
-        "counter.go",
-        "metric.go",
-        "opts.go",
-        "registry.go",
-        "util.go",
-        "version_parser.go",
-        "wrappers.go",
-    ],
+    srcs = ["util.go"],
     importpath = "k8s.io/kubernetes/pkg/util/metrics",
     deps = [
-        "//pkg/version:go_default_library",
-        "//staging/src/k8s.io/apimachinery/pkg/version:go_default_library",
         "//staging/src/k8s.io/client-go/util/flowcontrol:go_default_library",
-        "//vendor/github.com/blang/semver:go_default_library",
         "//vendor/github.com/prometheus/client_golang/prometheus:go_default_library",
-        "//vendor/github.com/prometheus/client_model/go:go_default_library",
-        "//vendor/k8s.io/klog:go_default_library",
     ],
 )
 
 go_test(
     name = "go_default_test",
-    srcs = [
-        "counter_test.go",
-        "registry_test.go",
-        "util_test.go",
-        "version_parser_test.go",
-    ],
+    srcs = ["util_test.go"],
     embed = [":go_default_library"],
     deps = [
-        "//staging/src/k8s.io/apimachinery/pkg/version:go_default_library",
         "//staging/src/k8s.io/client-go/util/flowcontrol:go_default_library",
-        "//vendor/github.com/blang/semver:go_default_library",
-        "//vendor/github.com/prometheus/client_golang/prometheus:go_default_library",
-        "//vendor/github.com/prometheus/common/expfmt:go_default_library",
-        "//vendor/github.com/stretchr/testify/assert:go_default_library",
     ],
 )
 
@@ -57,6 +34,9 @@ filegroup(
 
 filegroup(
     name = "all-srcs",
-    srcs = [":package-srcs"],
+    srcs = [
+        ":package-srcs",
+        "//pkg/util/metrics/framework:all-srcs",
+    ],
     tags = ["automanaged"],
 )
